apps/eonet/requirements.txt
functions-framework==3.*
requests>=2.25.0
google-cloud-storage>=2.0.0
requests>=2.25.0
